+++ javascripts/discourse/components/custom-footer.hbs
@@ -1,0 +1,67 @@
+{{#if @showFooter}}
+  <div class="wrap">
+    <div class="flexbox">
+      <div class="first-box">
+        <div class="heading">
+          {{this.mainHeading}}
+        </div>
+        <div class="blurb">
+          {{this.blurb}}
+        </div>
+      </div>
+      <div class="second-box">
+        <div class="links">
+          {{#each this.linkSections as |section|}}
+            <div class="list">
+              <span class={{section.className}} title={{section.title}}>
+                {{section.text}}
+              </span>
+
+              <ul>
+                {{#each section.childLinks as |link|}}
+                  <li class="footer-section-link-wrapper {{link.className}}">
+                    <a
+                      class="footer-section-link"
+                      title={{link.title}}
+                      href={{link.href}}
+                      target={{link.target}}
+                    >
+                      {{link.text}}
+                    </a>
+                  </li>
+                {{/each}}
+              </ul>
+            </div>
+          {{/each}}
+        </div>
+      </div>
+      <div class="third-box">
+        <div class="footer-links">
+          {{#each this.smallLinks as |link|}}
+            <a
+              class="small-link {{link.className}}"
+              title={{link.title}}
+              target={{link.target}}
+              href={{link.href}}
+            >
+              {{link.text}}
+            </a>
+          {{/each}}
+        </div>
+
+        <div class="social">
+          {{#each this.socialLinks as |link|}}
+            <a
+              class="social-link {{link.className}}"
+              title={{link.title}}
+              target={{link.target}}
+              href={{link.href}}
+            >
+              {{d-icon link.icon}}
+            </a>
+          {{/each}}
+        </div>
+      </div>
+    </div>
+  </div>
+{{/if}}
